Fix two temperature encoding defects found by bench dual-emit
Both were invisible on the car because its data never leaves the band real
driving produces. BENCH_IGNORE_PEC drives the same packing code with values at
and past the rails, which is where these live.

maxTemp was initialised with numeric_limits<float>::min(). For a floating-point
type that is the smallest positive normalised value, ~1.175e-38 -- not the most
negative, which is lowest(). For integer types min() *is* the most negative,
which is exactly what makes this misread. So the running maximum started at
approximately zero rather than at -infinity.

Any reading above freezing still won, so the car never showed it. But with the
whole pack below 0 C nothing ever exceeds the initialiser: TempMax reports
0.0 C and maxTemp_box stays at its invalid 255. That is a cold-weather bug on
the car, not only a bench artefact. minTemp used max() and was always right,
which is why TempMin agreed 38 of 38 in the same run where TempMax disagreed
38 of 38 -- the asymmetry in the results mirrors the asymmetry in the code.

Die temperatures were stored unsigned but have always been declared signed in
the DBC. A reading above 127 decoded negative -- 136 read as -120 -- and one
below 0 was lost on the conversion. The DBC is the correct side: the LTC6813's
die-temperature range is -40 to +125 C, which fits int8_t exactly and does not
fit uint8_t. Widened to int8_t end to end.

The conversion is now clamped before the cast. A float-to-integer conversion
whose result does not fit the destination is undefined behaviour, and a failed
read leaves raw_temp anywhere in 0..65535 -- which is precisely the case this
code is reached in when the chain is not answering.

The wire format is unchanged and the schema hash does not move: the signal was
already declared signed and is still eight bits of the same two's-complement
byte. Only the firmware's own storage was wrong. The CSV already cast to int
before printing, so it was never affected either way.

Verified by re-running the same capture: 8,039 comparisons with 77
disagreements before, 7,949 with none after. DieTemp_all 429/39 -> 448/0,
TempMax 0/38 -> 38/0.

+++ src/Telemetry.cpp
@@ -23,7 +23,7 @@ namespace {
  *   joules         J
  *   allVoltages    mV
  *   allTemperatures  degC as float
- *   dieTemps       whole degC, stored in a uint8_t
+ *   dieTemps       whole degC, stored in an int8_t (signed: the part's range is -40 to +125 C)
  *   heatsinktemp   whole degC (Main.cpp already divides the inverter's 0.1 degC by 10)
  */
 
@@ -171,7 +171,7 @@ void telemetry_emit_slow(const batterydata_t &d, uint32_t maxLatencyUs)
         for (uint8_t k = 0; k < 8; k++) {
             uint16_t idx = frame * 8 + k;
             if (idx < NUM_CHIPS) {
-                p[k] = d.dieTemps[idx];
+                p[k] = (uint8_t)d.dieTemps[idx];
             }
         }
         slcan_emit(TLM_ID_DIE_BASE + frame, p, 8, true);
